Moved format and spatial functions to separate modules
